I completed one main goal.
I displayed the first maze.

The maze is styled according to character.

My additions in the scripts.js file from line 55 to 77 where to
assign each character a td.

I also changed the level make up to be an array of 20 long strings.

In conclusion all the additions where towards the main goal of
displaying the first maze.